Making TwistReceived::ProcessMessage virtual
Use case : I want to write a specific ProcessMessage when a twist is received,  for example in my application, to modify the velocity of the four rotors of a quadrotor. 

I therefore thought about using a TwistReceiver  but what is done when a twist is received is actually hard coded in the twist receiver.  Therefore, there is a need to modify what is done by the TwistReceiver::ProcessMessage method. 

I see two possibilities :
1) making TwistReceiver   an abstract class   with ProcessMessage a virtual method  ; And then writting derived classes with their specific ProcessMessage
2) (the one implemented in this pull request)   Keep your "default" ProcessMessage   but add the virtual keywords to allow overriding it.