Adding texture filenames to ResourceFileURI list.
Some urdf files provide textures as parts of materials,
for example [1] uses

  <material name="Wheel_l">
    <texture filename="package://pr2_description/materials/textures/pr2_wheel_left.png" />
  </material>

These textures would not be requested, resulting in warnings during the
mesh import. This commit changes the ReadResourceFileUris such that it
does not only search for `mesh` nodes but also for `texture` nodes.
Further nodes can easily be added in the future.

[1] https://github.com/PR2/pr2_common/blob/kinetic-devel/pr2_description/urdf/materials.urdf.xacro