Update for openupm (#427)
* chore: Update name to com.siemens.ros-sharp

Currently all unity based extensions are in a folder called
Unity3D. This updates it to be named com.siemens.ros-sharp.

The purpose of this is so that we might be able to make an openupm
or npmjs package out of this that can be regularly published. This
avoids users of the Unity based extensions adding them in as
Assets, instead allowing them to add the extensions as an extenal
package. (see https://openupm.com/docs/ for more)

* chore: Add package.json

This adds a package json, so that we might be able to push this to
openupm as a distinct package.